Move mapif functions from char.c to mapif.c
